update CMSIS libarary, update main README.md

+++ tflite/micro/tools/cmsis/CMSIS/NN/Source/ActivationFunctions/arm_relu_q7.c
@@ -21,15 +21,15 @@
  * Title:        arm_relu_q7.c
  * Description:  Q7 version of ReLU
  *
- * $Date:        May 29, 2020
- * $Revision:    V.1.0.2
+ * $Date:        February 27, 2020
+ * $Revision:    V.1.0.1
  *
  * Target Processor:  Cortex-M cores
  *
  * -------------------------------------------------------------------- */
 
 #include "cmsis/CMSIS/DSP/Include/arm_math.h"
-#include "cmsis/CMSIS/NN/Include/arm_nnfunctions.h"
+#include "arm_nnfunctions.h"
 
 /**
  *  @ingroup groupNN
@@ -69,7 +69,7 @@ void arm_relu_q7(q7_t *data, uint16_t size)
         in = read_q7x4_ia(&input);
 
         /* extract the first bit */
-        buf = (int32_t)__ROR((uint32_t)in & 0x80808080, 7);
+        buf = __ROR(in & 0x80808080, 7);
 
         /* if MSB=1, mask will be 0xFF, 0x0 otherwise */
         mask = __QSUB8(0x00000000, buf);
